requirements.txt
ArchivesSnake==0.8.2
attrs==19.3.0
Babel==2.7.0
boltons==19.1.0
certifi==2019.9.11
chardet==3.0.4
clinner==1.12.3
colorlog==3.2.0
Django==2.2.6
djangorestframework==3.10.2
ElectronBonder==0.5
gitdb2==2.0.6
GitPython==3.0.3
health-check==3.4.1
idna==2.8
importlib-metadata==0.23
iso-639==0.4.5
jsonschema==3.1.1
more-itertools==7.2.0
multidict==4.5.2
objectpath==0.6.1
odin==1.4.0
psycopg2-binary==2.7.7
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
smmap2==2.0.5
sqlparse==0.3.0
structlog==19.1.0
urllib3==1.25.6
vcrpy==2.1.0
Wikidata==0.6.1
Wikipedia-API==0.5.2
wrapt==1.11.2
yarl==1.3.0
zipp==0.6.0
git+https://github.com/RockefellerArchiveCenter/asterism@v0.1#egg=asterism
